Refactor Delete ServicePuzzle interface Renamed DAO methods Renamed PuzzleService methods
