+++ puzzle-showdown/src/main/java/com/codecool/puzzleshowdown/controller/ControllerPuzzle.java
@@ -1,7 +1,7 @@
 package com.codecool.puzzleshowdown.controller;
 
 import com.codecool.puzzleshowdown.dto.puzzle.PuzzleDTO;
-import com.codecool.puzzleshowdown.service.ServicePuzzle;
+import com.codecool.puzzleshowdown.service.PuzzleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -13,25 +13,25 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class ControllerPuzzle {
 
-    private final ServicePuzzle servicePuzzle;
+    private final PuzzleService puzzleService;
 
     @Autowired
-    public ControllerPuzzle(ServicePuzzle servicePuzzle) {
-        this.servicePuzzle = servicePuzzle;
+    public ControllerPuzzle(PuzzleService puzzleService) {
+        this.puzzleService = puzzleService;
     }
 
     @GetMapping("/api/puzzle")
     public PuzzleDTO getPuzzle(){
-        return servicePuzzle.getPuzzle();
+        return puzzleService.getRandomPuzzle();
     }
 
     @GetMapping("/api/puzzle/valid/{puzzleId}/{move}/{step}")
     public String checkValidMove(@PathVariable String puzzleId, @PathVariable String move, @PathVariable int step){
-        return servicePuzzle.isValid(puzzleId, move, step);
+        return puzzleService.isValidStep(puzzleId, move, step);
     }
 
     @PatchMapping("/api/puzzle/{puzzleId}/popularity/{vote}")
     public void givePopularity(@PathVariable String puzzleId, @PathVariable int vote){
-        servicePuzzle.givePopularityToPuzzle(puzzleId, vote);
+        puzzleService.updatePopularity(puzzleId, vote);
     }
 }
