+++ puzzle-showdown/src/main/java/com/codecool/puzzleshowdown/dao/PuzzleDAO.java
@@ -3,7 +3,7 @@ package com.codecool.puzzleshowdown.dao;
 import com.codecool.puzzleshowdown.dao.model.Puzzle;
 
 public interface PuzzleDAO {
-    Puzzle getPuzzle();
-    void givePopularity(String puzzleId, int vote);
-    Puzzle checkValidMove(String puzzleId);
+    Puzzle getRandomPuzzle();
+    void updatePopularity(String puzzleId, int vote);
+    Puzzle getPuzzle(String puzzleId);
 }
