--- puzzle-showdown/src/main/java/com/codecool/puzzleshowdown/service/ServicePuzzle.java
@@ -1,13 +1,0 @@
-package com.codecool.puzzleshowdown.service;
-
-import com.codecool.puzzleshowdown.dto.puzzle.PuzzleDTO;
-import org.springframework.stereotype.Service;
-
-@Service
-public interface ServicePuzzle {
-    PuzzleDTO getPuzzle();
-
-    void givePopularityToPuzzle(String puzzleId, int vote);
-
-    String isValid(String puzzleId, String move, int step);
-}
